refactor: the file Parsing.hs was too big
As the file Parsing.hs was too big and started to be non readable it has
been split into multiple files.
The functions which are used for the parsing but could be used for
something else have been moved into their own file.
For instance StringOperations contains the functions that parse strings.
The file ReadOperations contains the functions to transform strings into
ints.
The .gitignore has been updated to contain test files.