feat(cluster): improve randome cluster generation
The cluster random generation was leading to clusters without pixels as
they were too far from the pixels in the space.
Instead of generating the centroids on 0-255 we now generate them
depending on the biggest and smallest values for each colors which makes
more sense and allows the program to be faster and more accurate.